+++ examples/xarm6/xarm6_grasp_place_env.py
@@ -63,7 +63,7 @@ class XArm6GraspPlaceEnv:
                 rendered_envs_idx=list(range(min(10, self.num_envs))),
             ),
             viewer_options=gs.options.ViewerOptions(
-                max_FPS=int(0.5 / self.ctrl_dt),
+                refresh_rate=int(0.5 / self.ctrl_dt),
                 camera_pos=(1.5, -1.5, 1.2),
                 camera_lookat=(0.3, 0.0, self.table_height + 0.2),
                 camera_fov=40,